moved UserLoginData and UserSubjectInfo to models, along with their tests and data
